python/pyproject.toml
[tool.poetry]
name = "oes-interview"
version = "0.1.0"
description = "OES Interview Python library"
authors = ["Wes <wes@anthronewengland.com>"]
license = "LGPL-3.0-or-later"
readme = "README.md"
packages = [{ include = "oes", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"

# Dependencies just to create/validate interview_old state
pynacl = "^1.5.0"
attrs = "^22.2.0"
cattrs = "^22.2.0"
orjson = "^3.8.8"

# Dependencies to work with the interview_old process
jinja2 = { version = "^3.1.2", optional = true }
pyparsing = { version = "^3.0.9", optional = true }
importlib-metadata = { version = "^6.0.0", optional = true }
loguru = { version = "^0.6.0", optional = true }
ruamel-yaml = { version = "^0.17.21", optional = true }
typing-extensions = "^4.7.1"

# TODO: proper versioning of this
oes-template = { git = "https://github.com/Open-Event-Systems/template.git", branch = "main", optional = true }
oes-hook = { git = "https://github.com/Open-Event-Systems/hook.git", branch = "main" }

# Dependencies to run the interview_old service
blacksheep = { version = "^1.2.12", optional = true }
typed-settings = { version = "^23.0.0", optional = true }
uvicorn = { version = "^0.20.0", extras = ["standard"], optional = true }
publicsuffixlist = "^0.10.0.20230617"
email-validator = "^2.0.0.post2"
openapi3 = "^1.8.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
httpx = { version = "^0.23.3" }
myst-parser = "^0.18.1"
sphinx = "^5.3.0"
sphinx-rtd-theme = "^1.1.1"
pytest-asyncio = "^0.21.0"

[tool.poetry.extras]
advanced = [
    "jinja2",
    "pyparsing",
    "importlib-metadata",
    "loguru",
    "ruamel-yaml",
    "oes-template",
]
server = [
    "jinja2",
    "pyparsing",
    "importlib-metadata",
    "loguru",
    "ruamel-yaml",
    "oes-template",
    "blacksheep",
    "typed-settings",
    "uvicorn"
]

[tool.mypy]
python_version = "3.9"
namespace_packages = true
explicit_package_bases = true
mypy_path = "$MYPY_CONFIG_FILE_DIR/src"

# Field entry points
[tool.poetry.plugins."oes.interview.field"]
text = "oes.interview.input.field_types.text:TextField"
button = "oes.interview.input.field_types.button:Button"
#email = "oes.interview.config.fields.email:EmailField"
#number = "oes.interview.config.fields.number:NumberField"
#bool = "oes.interview.config.fields.bool:BoolField"
#select = "oes.interview.config.fields.select:SelectField"
#date = "oes.interview.config.fields.date:DateField"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
